// frontend/src/api/generated/models/purchaseRequestPayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PurchaseRequestPayload {
  material_id: string;
  requested_qty: number | string;
  production_order_id?: string | null;
}