// frontend/src/api/generated/models/purchaseRequestResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PurchaseRequestResponse {
  id: string;
  pr_number: string;
  material_id: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  requested_qty: string;
  status: string;
  production_order_id: string | null;
}